+++ apps/chat/web/js/components/controllers/mainController.js
@@ -1,8 +1,30 @@
 var societyProChatControllers = 
-angular.module('societyProChatApp.controllers',['ngMaterial', 'societyProChatApp.controller2'])
+angular.module('societyProChatApp.controllers',['ngMaterial', 'ngAnimate', 'societyProChatApp.controller2'])
 
-.controller('mainController',['$scope','$http',function($scope,$http) {
-  $scope.token = "12345";
+.controller('mainController',['$scope','$http','$rootScope','$window',function($scope,$http,$rootScope,$window) {
+  
+  
+  var positionCommpanelScrollbar = function () {
+    var $cp = $('#sopro-collections-wrap');
+    if ($cp[0].scrollHeight > $cp.innerHeight()) {
+      $cp.perfectScrollbar();
+      $cp.perfectScrollbar('update');
+    } else {
+      if ($cp.hasClass("ps-container")) {
+        $cp.perfectScrollbar('destroy');
+        $cp.removeClass("ps-container")
+      }
+    }
+  }
+  setTimeout(function () {
+    positionCommpanelScrollbar();
+  }, 100);
+  var win = angular.element($window);
+  win.bind("resize",function(e){
+    positionCommpanelScrollbar();
+  });
+
+  $rootScope.token = "12345";
 
   $scope.roles = [
     {
@@ -19,16 +41,61 @@ angular.module('societyProChatApp.controllers',['ngMaterial', 'societyProChatApp
 
   $scope.channels = [];
   $scope.peers = [];
-  $scope.currentRole = {};
+  $rootScope.currentRole = {};
+  $scope.showCollectionsOverflow = null;
+  $scope.showSubscribersOverflow = null;
+
+  $(document).mousedown(function (e) {
+    console.log('mouseup listener');
+
+    var container1 = $("sopro-collections-dropdown");
+    var container2 = $("sopro-subscribers-dropdown");
+
+    // Determine which, if any, dropdowns to close based on a click event somewhere on the page
+    var isOpen1 = $scope.showCollectionsOverflow;
+    var isOpen2 = $scope.showSubscribersOverflow;
+
+    var clicked1 = container1.is(e.target);
+    var clicked1child = container1.has(e.target).length > 0;
+
+    var clicked2 = container2.is(e.target);
+    var clicked2child = container2.has(e.target).length > 0;
+
+
+    // if the target of the click isn't the container1...
+    // ... nor a descendant of the container1
+    // and it's already open...
+    if ( !clicked1
+      && !clicked1child
+      && isOpen1
+      && !clicked2
+      && !clicked2child
+    ){
+      $rootScope.$broadcast("collections.overflow.close", e.target);
+      $scope.showCollectionsOverflow = false;
+      $scope.$apply();
+    }
+
+
+    if (!clicked2
+      && !clicked2child
+      && isOpen2) 
+    {
+      $rootScope.$broadcast("subscribers.overflow.close", e.target);
+      $scope.showSubscribersOverflow = false;
+      $scope.$apply();
+    } 
+
+  });
 
   $scope.changeRole = function (role) {
-      $scope.currentRole = role;
+      $rootScope.currentRole = role;
 
       $http({
         method: 'GET',
         url: '/api/channels',
         headers: {
-         'token-auth': $scope.token
+         'token-auth': $rootScope.token
         },
         params : {
           role: role.id
@@ -40,6 +107,9 @@ angular.module('societyProChatApp.controllers',['ngMaterial', 'societyProChatApp
           //console.log(data);
           $scope.channels = data.channels;
           $scope.peers = data.peers;
+          setTimeout(function () {
+            positionCommpanelScrollbar();
+          }, 50);
         })
         .error(function(data, status, headers, config) {
           // called asynchronously if an error occurs
@@ -52,4 +122,18 @@ angular.module('societyProChatApp.controllers',['ngMaterial', 'societyProChatApp
 
 
   $scope.changeRole($scope.roles[0]);
+
+  // Open/Close Dropdown Logic:
+
+  $scope.$on("POO.click.collections", function ($event, data) {
+    $scope.showSubscribersOverflow = false;
+    $scope.showCollectionsOverflow = true;
+  });
+
+  $scope.$on("POO.click.subscribers", function ($event, data) {
+    $scope.showSubscribersOverflow = true;
+  });
+
+
+
 }]);